+++ CHANGELOG.rst
@@ -23,6 +23,7 @@ Added
 - Render detection function added.
 - CSV files saving for losses added in training.
 - Options to include the loss types in training.
+- Saving sample rendered images for truth and prediction.
 
 [0.1.11] - 2024-08-04
 ---------------------